% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overdispersion.R
\name{beta_info,overdispersion-method}
\alias{beta_info,overdispersion-method}
\title{Beta Information for Overdispersion}
\usage{
\S4method{beta_info}{overdispersion}(term, data)
}
\arguments{
\item{data}{data frame}

\item{object}{overdispersion object}

\item{...}{additional arguments}
}
\value{
NULL (overdispersion doesn't have beta parameters)
}
\description{
Extracts beta parameter information for overdispersion term.
}
