% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{beta_info}
\alias{beta_info}
\alias{beta_info,fpoly-method}
\alias{beta_info,linear-method}
\alias{beta_info,iwp-method}
\alias{beta_info,hiwp-method}
\alias{beta_info,iid-method}
\alias{beta_info,rpoly-method}
\alias{beta_info,hrpoly-method}
\title{Generic beta_info function}
\usage{
beta_info(term, data)

\S4method{beta_info}{fpoly}(term, data)

\S4method{beta_info}{linear}(term, data)

\S4method{beta_info}{iwp}(term)

\S4method{beta_info}{hiwp}(term)

\S4method{beta_info}{iid}(term)

\S4method{beta_info}{rpoly}(term)

\S4method{beta_info}{hrpoly}(term)
}
\arguments{
\item{term}{A model term object}

\item{data}{A data frame containing the data (not used for beta_info but kept for consistency with gamma_info)}
}
\value{
A data frame with beta (fixed effects) parameter information for the term, or NULL for terms without beta parameters
}
\description{
Generic beta_info function
}
