% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirichlet_multinom.R
\docType{class}
\name{dirichlet_multinom-class}
\alias{dirichlet_multinom-class}
\alias{dirichlet_multinom}
\alias{elgm_matrix,dirichlet_multinom-method}
\alias{design,dirichlet_multinom-method}
\alias{precision,dirichlet_multinom-method}
\alias{theta_info,dirichlet_multinom-method}
\alias{beta_info,dirichlet_multinom-method}
\alias{random_info,dirichlet_multinom-method}
\title{Dirichlet-multinomial response term}
\usage{
\S4method{elgm_matrix}{dirichlet_multinom}(term, data)

\S4method{design}{dirichlet_multinom}(term, data)

\S4method{precision}{dirichlet_multinom}(term, data)

\S4method{theta_info}{dirichlet_multinom}(term)

\S4method{beta_info}{dirichlet_multinom}(term, data)

\S4method{random_info}{dirichlet_multinom}(term, data)
}
\arguments{
\item{term}{A \code{dirichlet_multinom} term object.}

\item{data}{Analysis-ready data frame.}
}
\value{
Sparse \code{ngCMatrix} mapping observations to strata columns.
}
\description{
Model term for the case-crossover Dirichlet-multinomial log density
registered as \code{dirichlet_multinomial}.
}
\section{Functions}{
\itemize{
\item \code{elgm_matrix(dirichlet_multinom)}: Case-crossover stratum map for AD shards.

\item \code{design(dirichlet_multinom)}: Design method (not used).

\item \code{precision(dirichlet_multinom)}: Precision method (not used).

\item \code{theta_info(dirichlet_multinom)}: Theta info for overdispersion \code{tau}.

\item \code{beta_info(dirichlet_multinom)}: Beta info method (not used).

\item \code{random_info(dirichlet_multinom)}: Random info method (not used).

}}
\section{Slots}{

\describe{
\item{\code{by}}{Character vector of stratification variables for the case-crossover design.}
}}

\section{Slots (inherited from \code{model})}{

\describe{
\item{\code{ad_fun}}{Character scalar \code{"dirichlet_multinomial"}.}
\item{\code{ad_kind}}{Character scalar \code{"observations"}.}
\item{\code{package}}{Character scalar \code{"hpolcc"}.}
}
}

