% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{f}
\alias{f}
\title{Functions to Specify Terms of the Formula}
\usage{
f(x, model = "iid", ...)
}
\arguments{
\item{x}{A numeric vector representing the data to which the model is applied.}

\item{model}{A character string specifying the model of model term. Options include:
- `"fpoly"`: Fixed polynomial effect terms.
- `"rpoly"`: Random polynomial effect terms.
- `"hrpoly"`: Random hierarchical polynomial effect terms.
- `"iwp"`: Integrated Wiener process terms (random).
- `"hiwp"`: Hierarchical integrated Wiener process terms (random).}

\item{...}{Additional arguments passed to other methods or functions.}

\item{p}{A numeric value specifying the polynomial degree (default: varies by function).}

\item{ref_value}{A numeric value specifying the reference point for the basis functions.}

\item{knots}{A numeric vector specifying the knot locations for spline-based models.}

\item{range}{A numeric vector of length 2 specifying the range of the data (optional).}

\item{group_var}{A factor or grouping variable used to define hierarchical structures (optional).}

\item{include_global}{A logical value indicating whether to include a global component in hierarchical models (default: `TRUE`).}

\item{rpoly_p}{A numeric value specifying the degree for (random effects) polynomial components to add to the model as a separate term.}

\item{fpoly_p}{A numeric value specifying the degree for (fixed effects) polynomial components to add to the model as a separate term.}

\item{hrpoly_p}{A numeric value specifying the degree for hierarchical (random effects) polynomial components to add to the model as a separate term.}
}
\value{
A list or matrix, depending on the function, representing the design matrix, precision matrix, or parameter information for the specified model term.
}
\description{
This suite of functions implements various models (fixed and random effects), 
polynomials, integrated Wiener processes, and hierarchical extensions. The primary 
interface for these models is the `f()` function, which allows specification of 
model terms via the `model` argument.
}
\details{
The `f()` function serves as the primary entry point for specifying model terms. Internally, it dispatches to specific functions based on the `model` argument:
- `"fpoly"`: Calls the `fpoly()` function to specify fixed polynomial effect terms.
- `"rpoly"`: Calls the `rpoly()` function to specify random polynomial effect terms.
- `"hrpoly"`: Calls the `hrpoly()` function to specify hierarchical random polynomial effect terms.
- `"iwp"`: Calls the `iwp()` function to construct integrated Wiener process terms.
- `"hiwp"`: Calls the `hiwp()` function for hierarchical integrated Wiener process terms.
}
\examples{
# Example usage with f()
term <- f(x = 0:10, model = "iwp", ref_value = 5, knots = seq(0,10,2))

}
