% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpoly.R
\docType{class}
\name{fpoly-class}
\alias{fpoly-class}
\title{Fixed Polynomial Model Term}
\arguments{
\item{x}{Variable name}

\item{p}{Polynomial degree (default: 2)}

\item{ref_value}{Reference value for the polynomial}

\item{init}{Initial values for beta parameters}

\item{lower}{Lower bounds for beta parameters}

\item{upper}{Upper bounds for beta parameters}

\item{parscale}{Parameter scales for optimization}
}
\value{
A fpoly term object
}
\description{
Creates a fixed polynomial model term.
}
\examples{
# Create a fixed polynomial term
fpoly_term <- fpoly(x = "temp", p = 3, ref_value = 15)
}
