% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{gamma_info}
\alias{gamma_info}
\alias{gamma_info,iwp-method}
\alias{gamma_info,hiwp-method}
\alias{gamma_info,fpoly-method}
\alias{gamma_info,hrpoly-method}
\alias{gamma_info,rpoly-method}
\alias{gamma_info,iid-method}
\alias{gamma_info,linear-method}
\title{Generic gamma_info function}
\usage{
gamma_info(term, data)

\S4method{gamma_info}{iwp}(term, data)

\S4method{gamma_info}{hiwp}(term, data)

\S4method{gamma_info}{fpoly}(term, data)

\S4method{gamma_info}{hrpoly}(term, data)

\S4method{gamma_info}{rpoly}(term, data)

\S4method{gamma_info}{iid}(term, data)

\S4method{gamma_info}{linear}(term, data)
}
\arguments{
\item{term}{A model term object}

\item{data}{A data frame containing the data}
}
\value{
A data frame with gamma information for the term
}
\description{
Generic gamma_info function
}
