% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_ad_fun_raw_hpolcc}
\alias{get_ad_fun_raw_hpolcc}
\title{Build raw AD pack external pointer (hpolcc)}
\usage{
get_ad_fun_raw_hpolcc(data, config)
}
\arguments{
\item{data}{Model data list.}

\item{config}{Model configuration list.}
}
\value{
External pointer of class \code{adlaplace_handle_ptr}.
}
\description{
Build raw AD pack external pointer (hpolcc)
}
\keyword{internal}
