% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiwp.R
\docType{class}
\name{hiwp-class}
\alias{hiwp-class}
\title{Hierarchical Integrated Wiener Process Model Term}
\arguments{
\item{x}{Variable name}

\item{p}{Order of the integrated Wiener process (default: 2)}

\item{ref_value}{Reference value for the basis}

\item{knots}{Vector of knot locations}

\item{range}{Range of the data (optional)}

\item{by}{Grouping variable for hierarchical structure}

\item{init}{Initial values for theta parameters}

\item{lower}{Lower bounds for theta parameters}

\item{upper}{Upper bounds for theta parameters}

\item{parscale}{Parameter scales for optimization}

\item{boundary_is_random}{Whether boundary should be treated as random}

\item{include_poly}{Whether to include polynomial terms}

\item{include_global}{Whether to include global component}

\item{prefix}{Optional prefix for term names}
}
\value{
A list containing hiwp term object and optionally related terms
}
\description{
Creates a hierarchical integrated Wiener process (HIWP) model term.
}
