% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrpoly.R
\docType{class}
\name{hrpoly-class}
\alias{hrpoly-class}
\title{Hierarchical Random Polynomial Model Term}
\arguments{
\item{x}{Variable name}

\item{p}{Polynomial degree (default: 1)}

\item{ref_value}{Reference value for the polynomial}

\item{by}{Grouping variable for hierarchical structure}

\item{init}{Initial value for theta parameter}

\item{lower}{Lower bound for theta parameter}

\item{upper}{Upper bound for theta parameter}

\item{parscale}{Parameter scale for optimization}
}
\value{
A hrpoly term object
}
\description{
Creates a hierarchical random polynomial model term.
}
