% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_classes.R
\name{model_constructors}
\alias{model_constructors}
\title{Update model functions to return class objects}
\description{
These functions wrap the original model functions to return proper class objects
instead of lists.
}
