% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{precision}
\alias{precision}
\alias{precision,iwp-method}
\alias{precision,hiwp-method}
\alias{precision,rpoly-method}
\alias{precision,hrpoly-method}
\alias{precision,fpoly-method}
\alias{precision,iid-method}
\alias{precision,linear-method}
\title{Generic precision function}
\usage{
precision(term, data)

\S4method{precision}{iwp}(term, data)

\S4method{precision}{hiwp}(term, data)

\S4method{precision}{rpoly}(term, data)

\S4method{precision}{hrpoly}(term, data)

\S4method{precision}{fpoly}(term, data)

\S4method{precision}{iid}(term, data)

\S4method{precision}{linear}(term, data)
}
\arguments{
\item{term}{A model term object}

\item{data}{A data frame containing the data}
}
\value{
A precision matrix
}
\description{
Generic precision function
}
