% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overdispersion.R
\name{random_info,overdispersion-method}
\alias{random_info,overdispersion-method}
\title{Random Information for Overdispersion}
\usage{
\S4method{random_info}{overdispersion}(term, data)
}
\arguments{
\item{data}{data frame}

\item{object}{overdispersion object}

\item{...}{additional arguments}
}
\value{
NULL (overdispersion doesn't have random effects)
}
\description{
Extracts random effects information for overdispersion term.
}
