% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirichlet_multinom-reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{dirichlet_multinom}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{adlaplace}{\code{\link[adlaplace]{dirichlet_multinom}}}
}}

