% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_holidays.R
\name{remove_holidays}
\alias{remove_holidays}
\title{Remove Holidays from Data}
\usage{
remove_holidays(data, type = "rm_all")
}
\arguments{
\item{data}{A data frame containing a 'date' column.}

\item{type}{A character string specifying the type of holiday removal. Options are "rm_none", "rm_all".}
}
\value{
A data frame with holidays removed according to the specified type.
}
\description{
This function removes holidays from a dataset based on the specified type.
}
