% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsiwp.R
\docType{class}
\name{rsiwp-class}
\alias{rsiwp-class}
\alias{design,rsiwp-method}
\alias{precision,rsiwp-method}
\alias{theta_info,rsiwp-method}
\alias{beta_info,rsiwp-method}
\alias{random_info,rsiwp-method}
\title{Random Slope Integrated Wiener Process Term}
\usage{
\S4method{design}{rsiwp}(term, data)

\S4method{precision}{rsiwp}(term, data)

\S4method{theta_info}{rsiwp}(term)

\S4method{beta_info}{rsiwp}(term)

\S4method{random_info}{rsiwp}(term, data)
}
\arguments{
\item{term}{An rsiwp term object}

\item{data}{A data frame containing the term variable}
}
\description{
Creates and manages random slope integrated Wiener process (RSIWP) model terms.
}
\section{Functions}{
\itemize{
\item \code{design(rsiwp)}: Creates design matrix for RSIWP term

\item \code{precision(rsiwp)}: Creates precision matrix for RSIWP term

\item \code{theta_info(rsiwp)}: Extracts theta parameter information for RSIWP term

\item \code{beta_info(rsiwp)}: Extracts beta parameter information for RSIWP term

\item \code{random_info(rsiwp)}: Extracts random effects information for RSIWP term

}}
\section{Methods}{

The following methods are available for `rsiwp` objects:
\describe{
  \item{\code{design(term, data)}}{Creates design matrix for RSIWP term}
  \item{\code{precision(term, data)}}{Creates precision matrix for RSIWP term}
  \item{\code{theta_info(term)}}{Extracts theta parameter information}
  \item{\code{beta_info(term, data)}}{Extracts beta parameter information}
  \item{\code{random_info(term, data)}}{Extracts random effects information}
}
}

