% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsrpoly.R
\docType{class}
\name{rsrpoly-class}
\alias{rsrpoly-class}
\title{Random Polynomial Model for Random Slope}
\arguments{
\item{x}{Variable name}

\item{mult}{covariate to be mulitplied}

\item{p}{Polynomial degree (default: 2)}

\item{ref_value}{Reference value for the polynomial}

\item{ref_mult}{Reference value for the covariate}

\item{sd}{Standard deviation for random effects}
}
\value{
A rsrpoly term object
}
\description{
Creates a random polynomial model term for use in a random slope model.
}
