% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{theta_info}
\alias{theta_info}
\alias{theta_info,iwp-method}
\alias{theta_info,hiwp-method}
\alias{theta_info,iid-method}
\alias{theta_info,fpoly-method}
\alias{theta_info,rpoly-method}
\alias{theta_info,hrpoly-method}
\alias{theta_info,linear-method}
\title{Generic theta_info function}
\usage{
theta_info(term)

\S4method{theta_info}{iwp}(term)

\S4method{theta_info}{hiwp}(term)

\S4method{theta_info}{iid}(term)

\S4method{theta_info}{fpoly}(term)

\S4method{theta_info}{rpoly}(term)

\S4method{theta_info}{hrpoly}(term)

\S4method{theta_info}{linear}(term)
}
\arguments{
\item{term}{A model term object}

\item{theta_info}{Additional theta information (not currently used but kept for compatibility)}
}
\value{
A data frame with parameter information for the term
}
\description{
Generic theta_info function
}
