% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{get_new_xa}
\alias{get_new_xa}
\title{Post-fit design matrix constructor}
\usage{
get_new_xa(terms, df)
}
\arguments{
\item{terms}{The \code{terms} component from an \code{hnlm()} fit.}

\item{df}{A data frame with the columns required to rebuild the design
matrices.}
}
\value{
A list containing the fixed-effect matrix \code{X} and random-
effect matrix \code{A}.
}
\description{
Replicates the loop in \code{hnlm()} that constructs the design matrices,
without modifying \code{terms} or creating the associated precision
matrices.
}
