% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hnlm.R
\name{hnlm}
\alias{hnlm}
\title{Fit (some) hierarchical non-linear models (that we like)}
\usage{
hnlm(
  formula,
  data,
  cc_design = c(),
  config = list(dirichlet = TRUE, boundary_is_random = FALSE, transform_theta = TRUE),
  control = list(fnscale = -1, trace = 3, REPORT = 1, maxit = 1000),
  control_inner = list(report.level = 0),
  for_dev = FALSE,
  ...
)
}
\arguments{
\item{formula}{A formula object specifying the model to be fitted.}

\item{data}{A data frame containing the variables specified in the
formula and any additional variables required for the model.}

\item{cc_design}{An object specifying the case-crossover design,
including stratification and time variables. Defaults to the output
of \code{ccDesign()}.}

\item{for_dev}{Logical; if \code{TRUE}, the function returns intermediate
objects for development purposes. Defaults to \code{FALSE}.}

\item{dirichlet}{If \code{TRUE}, fit a dirichlet-multinomial model with a
time-within-strata gamma random effect; otherwise fit a multinomial
model.}

\item{verbose}{logical, if \code{TRUE} occasional information is printed}
}
\value{
A list containing the fitted TMB object, the formula, terms
used in the model, the case-crossover design, and information about
the gamma and theta parameters.
}
\description{
This function fits a hierarchical model using the specified formula
and data, incorporating case-crossover designs, fixed and random
effects, and precision matrices for random effects. It allows for
flexible inclusion of stratification variables, time variables, and
complex random effect structures.
}
\details{
The function handles fixed effects, random effects, and
their associated precision matrices. It also optimizes the model
using TMB with options for additional preprocessing and handling
specific random effect structures.
}
\examples{
# See vignette for basic usage

}
